+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/attack_details/main/hooks/use_attack_entities_counts.test.ts
@@ -8,11 +8,11 @@
 import { renderHook } from '@testing-library/react';
 import type { DataTableRecord } from '@kbn/discover-utils';
 import { useAttackEntitiesCounts } from './use_attack_entities_counts';
-import { useOriginalAlertIds } from './use_original_alert_ids';
+import { useHeaderData } from './use_header_data';
 import { useQueryAlerts } from '../../../../detections/containers/detection_engine/alerts/use_query';
 
-jest.mock('./use_original_alert_ids', () => ({
-  useOriginalAlertIds: jest.fn(),
+jest.mock('./use_header_data', () => ({
+  useHeaderData: jest.fn(),
 }));
 
 jest.mock('../../../../detections/containers/detection_engine/alerts/use_query', () => ({
@@ -25,13 +25,16 @@ const hit: DataTableRecord = {
   flattened: {},
 };
 
+const mockHeaderData = (originalAlertIds: string[]) =>
+  ({ originalAlertIds } as unknown as ReturnType<typeof useHeaderData>);
+
 describe('useAttackEntitiesCounts', () => {
-  const mockUseOriginalAlertIds = jest.mocked(useOriginalAlertIds);
+  const mockUseHeaderData = jest.mocked(useHeaderData);
   const mockUseQueryAlerts = jest.mocked(useQueryAlerts);
 
   beforeEach(() => {
     jest.clearAllMocks();
-    mockUseOriginalAlertIds.mockReturnValue([]);
+    mockUseHeaderData.mockReturnValue(mockHeaderData([]));
     mockUseQueryAlerts.mockReturnValue({
       loading: false,
       data: null,
@@ -43,7 +46,7 @@ describe('useAttackEntitiesCounts', () => {
   });
 
   it('returns zero counts and skips query when alertIds is empty', () => {
-    mockUseOriginalAlertIds.mockReturnValue([]);
+    mockUseHeaderData.mockReturnValue(mockHeaderData([]));
 
     const { result } = renderHook(() => useAttackEntitiesCounts(hit));
 
@@ -58,7 +61,7 @@ describe('useAttackEntitiesCounts', () => {
   });
 
   it('passes query with ids filter and cardinality aggs when alertIds exist', () => {
-    mockUseOriginalAlertIds.mockReturnValue(['id1', 'id2']);
+    mockUseHeaderData.mockReturnValue(mockHeaderData(['id1', 'id2']));
 
     renderHook(() => useAttackEntitiesCounts(hit));
 
@@ -78,7 +81,7 @@ describe('useAttackEntitiesCounts', () => {
   });
 
   it('parses relatedUsers and relatedHosts from aggregations', () => {
-    mockUseOriginalAlertIds.mockReturnValue(['id1']);
+    mockUseHeaderData.mockReturnValue(mockHeaderData(['id1']));
     mockUseQueryAlerts.mockReturnValue({
       loading: false,
       data: {
@@ -106,7 +109,7 @@ describe('useAttackEntitiesCounts', () => {
   });
 
   it('returns zero when aggregations are missing', () => {
-    mockUseOriginalAlertIds.mockReturnValue(['id1']);
+    mockUseHeaderData.mockReturnValue(mockHeaderData(['id1']));
     mockUseQueryAlerts.mockReturnValue({
       loading: false,
       data: {
