+++ x-pack/solutions/security/plugins/security_solution/public/timelines/components/timeline/unified_components/data_table/index.tsx
@@ -59,12 +59,7 @@ import { getTimelineRowTypeIndicator } from './get_row_indicator';
 import { isAttackDiscoveryRow } from './is_attack_discovery_row';
 import { DocumentFlyoutWrapper } from '../../../../../flyout_v2/document/main/document_flyout_wrapper';
 import { flyoutProviders } from '../../../../../flyout_v2/shared/components/flyout_provider';
-import {
-  defaultToolsFlyoutProperties,
-  useDefaultDocumentFlyoutProperties,
-} from '../../../../../flyout_v2/shared/hooks/use_default_flyout_properties';
-import { AttackDetails } from '../../../../../flyout_v2/attack_details/main';
-import { NotesDetails } from '../../../../../flyout_v2/shared/tools/notes';
+import { useDefaultDocumentFlyoutProperties } from '../../../../../flyout_v2/shared/hooks/use_default_flyout_properties';
 
 const DataGridMemoized = React.memo(UnifiedDataTable);
 
@@ -193,66 +188,26 @@ export const TimelineDataTableComponent: React.FC<DataTableProps> = memo(
     const handleOnEventDetailPanelOpened = useCallback(
       (eventData: DataTableRecord & TimelineItem) => {
         if (newFlyoutSystemEnabled) {
-          const isAttackRow = isAttackDiscoveryRow(eventData);
-
-          if (isAttackRow) {
-            const onShowNotes = () => {
-              overlays.openSystemFlyout(
-                flyoutProviders({
-                  services,
-                  store,
-                  history,
-                  children: <NotesDetails hit={eventData} />,
-                }),
-                {
-                  ...defaultToolsFlyoutProperties,
-                  historyKey: documentFlyoutHistoryKey,
-                }
-              );
-            };
-
-            overlays.openSystemFlyout(
-              flyoutProviders({
-                services,
-                store,
-                history,
-                children: (
-                  <AttackDetails
-                    hit={eventData}
-                    onShowNotes={onShowNotes}
-                    renderCellActions={cellActionRenderer}
-                    onAlertUpdated={refetch}
-                  />
-                ),
-              }),
-              {
-                ...defaultFlyoutProperties,
-                historyKey: documentFlyoutHistoryKey,
-                session: 'start',
-              }
-            );
-          } else {
-            overlays.openSystemFlyout(
-              flyoutProviders({
-                services,
-                store,
-                history,
-                children: (
-                  <DocumentFlyoutWrapper
-                    documentId={eventData._id}
-                    indexName={eventData.ecs._index}
-                    renderCellActions={cellActionRenderer}
-                    onAlertUpdated={refetch}
-                  />
-                ),
-              }),
-              {
-                ...defaultFlyoutProperties,
-                historyKey: documentFlyoutHistoryKey,
-                session: 'start',
-              }
-            );
-          }
+          overlays.openSystemFlyout(
+            flyoutProviders({
+              services,
+              store,
+              history,
+              children: (
+                <DocumentFlyoutWrapper
+                  documentId={eventData._id}
+                  indexName={eventData.ecs._index}
+                  renderCellActions={cellActionRenderer}
+                  onAlertUpdated={refetch}
+                />
+              ),
+            }),
+            {
+              ...defaultFlyoutProperties,
+              historyKey: documentFlyoutHistoryKey,
+              session: 'start',
+            }
+          );
         } else {
           const isAttackRow = isAttackDiscoveryRow(eventData);
           const indexName = eventData.ecs._index ?? '';
